backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
pandas-ta-classic
numpy
requests
beautifulsoup4
google-generativeai
anthropic
